[amber] Change initialization process
This change allows us to start the daemon without any sources.
This also remove the DaemonProvider structure in favor of always
providing the daemon directly. The FIDL server is still given a
way to kick the TUF client initialization logic.

Retries are added to the amber_ctl command since it is now much
more likely a request might fail if a source has just been added
or initialized.

Change-Id: I87ef93fd08e4f013efe3ddb5929fbb9dd457b206