[debugger] Improve symbol indexing.
Fixes a bug where functions with declarations were only indexed in the
first compile unit. The offset computation was incorrect. A test is
added for this case.

Improves indexing performance on large symbol files (I used Linux Chrome
which is 6.6GB) by a little over 2x. This is done by avoiding decoding
DWARF DIEs that we know can't contribute. I toyed around with how the
DIEs are decoded but it seems to not matter: just going through the
items in it is the slow part.

Change-Id: If5a4b8899bc38886e9382971bddaeadd935a43b4