[media_player] ensure decoders don't request packets after eos
Decoder nodes were recently modified to run decoders on their own
threads rather than on the graph's main thread. In making this
change, a bug was introduced in which decoder nodes may request
input packets after they receive an input packet marked end-of-
stream. Nodes aren't permitted to do this, and the demux DCHECKs
as a consequence.

This CL adds the additional state required to reflect that an
end-of-stream input packet has been encountered. This state has
been merged with the 'input packet requested' binary state so
it can easily be managed with a single atomic value.

Change-Id: I3fb491f589e8ac8f34da90198195e28666153b5e