[media_player] close hole in stage NeedsUpdate logic
The Update method in a stage responds to changes in supply (the
availability of packets at inputs) and demand (the requests for
packets at outputs). It also deals with the production of packets
by the node it hosts.

The need to run the Update method is signalled by calling the
NeedsUpdate method. NeedsUpdate posts a task that runs a method
called UpdateUntilDone, which calls Update repeatedly until all
state changes have been handled.

NeedsUpdate and UpdateUntilDone coordinate using an atomic counter.
A counter value of 0 indicates that UpdateUntilDone isn't running
and hasn't been queued. NeedsUpdate increments the counter atomically
and posts a call to UpdateUntilDone if the counter was 0 before the
increment. UpdateUntilDone sets the counter to 1 before calling
Update. After Update returns, UpdateUntilDone performs a compare
exchange on the counter, replacing the 1 with a 0 and terminating
if the exchange occurs. It loops around and calls Update again if
the counter wasn't 1. This ensures that Update runs again if
NeedsUpdate was called while Update was running.

NeedsUpdate was setting the counter to 2 if the counter wasn't zero
prior to the increment. There was no good reason to do this, and it
created opportunities for the update scheme to go wrong. If the
counter wasn't zero prior to the increment, and UpdateUntilDone then
set the counter to 1, ran update and did its compare/exchange before
NeedsUpdate set the counter to 2, the counter would end up with the
value 2 even though UpdateUntilDone had stopped looping. The stage
would essentially be stuck thereafter.

This CL removes the code in NeedsUpdate that sets the counter to 2.

Change-Id: I7192442ac94991005e43e6b50c0c81dc3e497611