[modular] Back versions for JSON-XDR filters
The XdrRead() call now takes instead of a single Xdr filter an array
of them. It tries to read using the first filter from the list first.
If that fails, instead of failing altogether, it descends the list
and tries to read using the remaining filters from the list, until
one succeeds or the end of the filter list is reached, in which case
it still fails.

Once a filter is changed, the old version can be retained and added to
the filter versions list. It is idiomatic to give the filter version list
the same name as a simple filter function, and call the individual
filter function versions with a suffix _v1, _v2 etc. to indicate the
version.

XdrWrite() is changed to take the same filter list as XdrRead, so
symmetry of read and write invocation is preserved. However, write
will only ever use the first filter on the list, i.e. it only ever
writes the latest version.

Added a unittest case for reading with versioned filters.

Introduce backwards compatibility filters for XdrStoryData and
XdrDeviceData, which got changed during the FIDL2 conversion with
provisioning for backwards compatibility, which gave rise to the
JIRA issues mentioned below. The back versions can read versions
used during the transition. The newest version can read current
data and data from before the FIDL2 conversion.

In order for the changed formats to work, we need XDR support for
fidl::Array. Added this to json_xdr.h, and a corresponding unittest.

Also change all existing XDR filters to filter version lists, and
replace existing read error handlers with corresponding back versions.

Collateral cleanup:

* Moved shared Xdr filter function definitions into their own
  .cc files, rather than to inline them in an .h file or duplicate
  them in multiple .cc files.

* Applied naming convention to name the Xdr filter function and
  version list directly after the data type it applies to.

TODO:

What we still need in a followup CL is support for regression testing
in integration tests, where we read ledger data from old versions and
verify the story runner picks up the right state.

Also possible is to allow filter lists not just in XdrRead()/XdrWrite(),
i.e. at the top level, but actually down at the Field() call.

MI4-999 #done
MI4-994 #done
MI4-995 #done

Change-Id: I07076dd3d9218800b69c07488c19a5b68e9b7a90