[zbi] Introduce libzbi
Introduce libzbi for ZBI manipulation from Zircon
userspace or kernel space.

There are a number of places in the system where we do
arbitrary ZBI manipulation:
 + Iterating over an image
 + Appending items to an image
 + Extracting items from an image
and it would be nice to have a single implementation
for each of these routines.

For now, this implementation provides the general
shape and layout of libzbi and ports the ARM64
platform's initialization routines to use libzbi.

If all looks well, I can continue to port x86 too.

Change-Id: I5e2825cc3e79ec12c0d408bc38e779102c75698b