[debugger] Symbol lookup by module.
Adds a ModuleSymbols class to store the symbol information for a given
module. This builds and holds the existing index class for the module
functions.

Adds a start of the symbol object model, introducing classes for a few
symbol-related things.

Adds querying capabilities to convert an address to a Location, and to
look up a function name and convert to an address.

clang-format update some random things in command_utils.cc

Change-Id: I03f10e5240651de85a4e3afa47a3569792637ef1