[kernel][mp] add new mp_interrupt() call to be used by the hypervisor
In some cases the hypervisor just needs to interrupt another cpu for the
purposes of getting it to vmexit. Add another specific mp_* call to just
do this. Allows for simplifying the mp_reschedule() call by removing a
now unused flag.
Remove a layer of indirection of the ipi interrupts which aren't useful.

Change-Id: Ib54fc8683338b6f9a8efeeb0e7478a0a5cdab095