[fidl][go] Only generate imports for used dependencies
This change ensures that only imports for dependencies which are used
are actually generated, otherwise transitive dependencies get imports
generated for them resulting in build failures in the generated code.

This change also cleans up the logic a bit, and merges the dependencies
on syscall/zx and fidl/bindings into the same logic.

FIDL-193

Change-Id: Ic84307eb982b095ae342bff7a401be5b6288e025