[crypto][zxcrypt] Remove Bytes::Increment
This CL removes the method was being used for incrementing the IV used
by AEAD and Cipher.  As written, it was too general purpose and slow in
order to be constant time.  In its place, AEAD and Cipher now represent
their IVs as an array of zx_off_ts, and only increment the first one.
this is much quicker (and still safe) as it is only an add and a store.
This limits both of them to 2^64 cryptographic operations, but in
practice this likely won't be a concern.

Change-Id: I9ed707a2aa71aae619906af5918f3d8e5451b3d5