[json_xdr] Add explicit versions to all XDR filters.
Versioning was there before, but with explicit version numbers
data in the ledger become more intelligible, and we can express
upgrades in the encoding of the value of a field without having
to change its name or its type.

We use XDR filters to create JSON in places that is not persisted,
specifically in context topics. There we don't write versions for
now, since version skew cannot happen.

Change-Id: Ifb2cd14bc291da7c9250ab5521fc49c4d6cfb000