[threads] Convert suspend to suspend_token calls.
Adds documentation for the new zx_task_suspend_token() and
zx_task_resume_from_exception() calls. Marks the existing calls as
deprecated in the documentation.

Fixes a bug in the debugger.cpp test where the test looped to wait for a
register to be the correct value, but never resumed the thread which
would allow this to happen.

A reasonable amount of code in the threads.cpp test code could be
deleted since it was trying to emulate the features of the suspend token
(automatic resumption on crash).

Adds zx::suspend_token wrapper object, and a creation function for it
zx::task::suspend(suspend_token*).

ZX-1072

Change-Id: Ife4e7e637d4f67564f891b7311dbcad61b9c7064