[benchmarks] Convert Thread/CreateAndJoin test to use the perftest library
We drop the use of PauseTiming/ResumeTiming() around the
zx_handle_close() calls.  The cost of those is trivial in comparison
with the rest of the test, and we measure the cost of
zx_handle_close() in other tests, so it's not worth making this into a
multi-step test.

TO-651

Change-Id: I65c0aa4b3b802514cd44e924590fa27647dba610