[suggestion_engine] Remove custom action delegations
CustomAction delegations used to let agents lazily resolve actions to
be taken after a suggestion was selected. This functionality has not
been needed for a few months, was broken at some point, and the broken
version seems to expose a stale pointer issue elsewhere. Removing it to
stop the crashing while we investigate further.

MI4-1005 #comment

Change-Id: Ia9a3944a85a4dfca292f46021fe046bb5ba4ef92