[debugger] Breakpoint improvements.
Actual functionality is unchanged.

Adds the concept of a symbolic breakpoint. Separate out the breakpoint
settings from the Breakpoint object so that all settings can be
atomically set at once (previously there was an awkward "Commit"
function that wouldn't work with symbolic breakpoints).

Added the concept of a "none" stop mode. None of the stop modes are
hooked up yet though. Created a new enum for stop mode in the client
API. I think exposing debug_ipc types in the client API is ugly and I
plan to reduce this.

Converted a few of the changed functions from a loop to a switch
statement. This allows the compiler to warn if an enum case isn't
handled.

Change-Id: Ife9fadadb98b16f4b43a8f983522c698f8943948