[media_player] decoder wrapper redesign
The decoders, which were recently overhauled to run on worker
threads, were working well *almost* all the time, but obscure races
continued to plague them. The bulk of this code is in
ffmpeg_decoder_base.*. This CL overhauls the decoders yet again to
simplify their operation.

The previous approach was to share a number of fields between the
main thread and worker thread in FfmpegDecoderBase in order to move
input packets from worker to main and output packets from main to
worker. Demand likewise flowed through the worker thread. This
approach worked, but had issues with obscure race conditions.

The new approach is to avoid sharing fields between the two threads,
manage all packet supply and demand on the main thread, and simplify
the worker. In addition, the specifics of ffmpeg decoding have been
factored away from the main/worker coordination by the introduction
of the SoftwareDecoder base class, which handles the latter. This
allows FfmpegDecoderBase to operate much as it did when it operated
synchronously.

SoftwareDecoder uses the worker thread as a one-at-a-time decoder of
input packets, and all main/worker coordination is done via posted
tasks. The main thread posts a task to the worker thread to process
an input packet. The worker completely processes the input packet in
that task, posting to the main thread to deliver any output packets
generated and to signal that input packet has been completely
processed. There is a worker thread Flush call for bookkeeping.

A mutex is used to synchronize access to a ValueTracker, which is
used exclusively for the introspection report.

Change-Id: I0477a306635d0075007e8d0a1630c562827026c5