[vmo-utils] Adding VmoPool
VMO Pools are collections of VMOs that are used together
and share similiar properties.  The VMO pool is intended to be used
by a content producer, as all VMOs in the pool are automatically
mapped to a VMAR.  The VMO Pool adds lifecyle management as well,
by keeping track of which vmos are 'locked'.

VMO Pools are intended to act as one backing for BufferCollections.

Change-Id: Icd998e0eb4b00c0192f5e143de07194a1e1f3930