[svchost] Whitelist deprecated service access
Some tests depend on being able to access system-level services in the
test environment. This CL adds a whitelist of those services so we can
systematically remove these dependencies.

The whitelist in this CL is overly broad, but once this CL cycles
through the bots, we'll be able to grep the logs and reduce it
substantially.

Change-Id: If144c00895a8408fefabbd78783a3e71fd62e5c1