[zircon][amlogic][sd-emmc] Add PIO mode support
The SDIO instance PORTA of the amlogic controller does not support DMA.
This change adds PIO support for the driver. The use_dma flag is used by the amlogic
controller for command submission as well as data as compared to the
sdhci on x64. In PIO mode, the controller reads the descriptors from the SRAM region of
the mmio space and data from the ping-pong buffers. For SDHCI on X64, the use_dma
flag is ignored during command submission and used only during data.
While in sdmmc, this changeset also fixes the issue of unmap when sdmmc
does not use dma.

Change-Id: I8d6dc95acb6405e1f89511376a68a044bbf0ece4