[libzbi] Add append section function
Support a mechanism to append ZBI sections to the
end of the ZBI in libzbi.

Currently there is a kernel only mechanism for
performing this action against bootdata
containers.

This mechanism will be removed and replaced with
the libzbi method.
There are a number of implementations of append
section in userspace that will also be replaced
by the libzbi method.

Change-Id: Iceb89334dcd75df3a42f77ebfa841ce95fd8190d