[zircon][hid-parser] handle 32-bit usages
Turns out we have a hid usage in the vendor-defined range
which is > uint16_max. See dev-tbmc.cpp

Now that the hid parser is live in Mozart, this generates "error"
spew in the system logs as the parser did not expect to deal with
this.

For usage ranges, we still keep the 16bit check, out of well
founded fear.

- The fix
- Two new tests

ZX-2180 #comment progress

Change-Id: I23e000d8aee2188fe50024c6c47171dee4af3cb0