[debugger] Move symbols to separate target
Moves the *symbols* files in zxdb/client/ to the symbols subdirectory.

Severs the dependencies from the symbols directory to the client
directory and makes it a separate target. This allows the symbol
functionality to be unittested in isolation from the rest of the
debugger.

The symbols implementation relied on two very low-level sources in
client/ so these were separated into a new small library to be shared.

There should be no functionality change.

Change-Id: I16df92c8f49008b1fcee0c704cd912bbf8bca146