[netstack] Add DhcpRequest Handling Logic
This change implements the message handling logic for DHCP Request
messages sent by a client in a Selecting, InitReboot, or Renewing state.
(Message handling for Rebinding clients has been deferred until lease
timeouts are implemented.) Additionally, this change adds a ClientState
enum and a handful of helper functions, both of which may be moved
elsewhere as dhcpd continues to be iterated upon.

test: passes unit tests

NET-752 #progress

Change-Id: Ic730cf9675b95303e654ebc9b4da292a6c72304f