[packages][magma] Revamp packaging
Currently the magma package includes intel-gen on x64
and arm-mali on arm64.  Now we have another arm64 target
with a different gpu, vsl-gc.

Create separate magma packages for each gpu, so that
eventually we can include just the one for a given build.

We can't yet specify what package to include, so include
them all for now.

Change-Id: I6d2b99b068ea29e3bafefd1ad66d49a073b9ccab