[x86] Unify FP fxsave area initialization.
The fxsave instruction is older and only supports the legacy save area.
Previously we would initialize the default save area by just calling
fxsave, assuming the x87 state on boot is the default value.

But since xsave and fxsave share the same legacy area, we can unify the
initialization and not depend on the initial register state.

Provide separate functions to initialize the FP and SSE state.
Previously the code was 0-initializing the FPU state which will not
count as being in the "initial configuration" for optimizations, and may
be surprising to applications (our apps don't use x87 so we never
noticed).

The initial state functions set everything in that state component. As
currently used this is unnecessary since the initial xsave area is
0-initialized, but these will be used in the future for reading
components out, and we need the ability to get the default if the
processor indicates it didn't write anything (due to being in the
initial state).

Clarify FTW field in the legacy xsave area. This is confusingly
different from the FTW register.

Change-Id: I688615435be46a3454f20af70a3d837c45eeccb6