[dev][intel-hda][dsp] Fix some of the stream channel handlers.
Use the utility base class's defaults for handling requests to report plug
detection capabilities/state and for settting/getting gain instead of simply
returning an uninitialized structure.

Also, do not report support for 1 channel audio on the supported formats list as
the code is not currently ready to handle that.

ZX-1538 #comment Fix at least on issue which could cause unreliability

Change-Id: Ia083c854f145d214e75bf4a3599fe11780ba4eec