[amber] Load config from a file
This switches the amber from hardcoding the localhost configuration
into loading from one or more files. This will make it easier to bake
in alternative repositories down the road. This patch creates a
directory, /pkg/data/sources, which can contain any number of json
source definitions. The initial one just describes the localhost
definition we had before.

One thing to note with this patch is it allows a blob repository
to be specified when creating a source. Technically speaking,
TUF sources don't need to provide for blob serving, so we might
want to provide a separate API for registering blob repos down
the road.

Change-Id: I77d9f70b2f164e2fd4de283afeb2118ccb12d8e5