[fit] Add fit::defer.
fit::defer is a replacement for fbl::AutoCall and fxl::AutoCall.
It behaves the same way but it has a default constructor and it
eagerly destroys its target after it is canceled or called much
like if the target was held in an std::optional (which we don't
have in C++14 unfortunately).

Bug: ZX-2222
Test: fit-test
Change-Id: I24f0004d468377c40925288929660aa9cc3705a6