[media_player] make FakeAudioRenderer operate in real time
This CL updates the FakeAudioRenderer used in media_player_tests so
it retires packets in real time according to the timeline established
by the Play method of AudioRendererV2. This is part of a general
upgrade of the CQ test to make it more like real operation and to
expand its coverage.

Change-Id: I8cf0165a8b24b7038394757cc31d2ef65b643a4e