[zircon][amlogic][sdmmc] Reduce to even lower frequency when tuning fails
The new vim2 boards return CRC errors during read operations at the
existing frequency. The tuning process actually tries to figure out the
optimum frequency by lowering the frequency and resending the tuning
command. However at times the tuning commands succeed but the read
command fails with that frequency. This changeset descreases the
frequency more when a tuning failure occurs so that the read commands
that follow succeed.

Change-Id: Iff96c3e8d5cda303e293be5f295dd616ee65864b