[wlanstack2] Separate module for watching /dev
Create streams of newly opened phy and iface devices. Instead of
listening for removal events in /dev, rely on the channels being
closed when device unbinds. Introduce a wlantap-based test for
watching wlanphy devices. Be careful to use select() over join()
when necessary.

This is the second part of the plan to take apart DeviceManager.

NET-923
NET-779

Change-Id: I07ae3b3ef276ff3ce815b91dbeb5a3032a4e59fc