[user_shell] Indicate ongoing story activity to user shell.
ModuleContext now has a method Active(). If a module has ongoing
activity, it is expected to call this method. Not calling this
method indicates that there is no ongoing activity, even though
the module continues to display UI and awaits user input etc.

StoryProvider now has a method WatchActivity() that lets the user
shell register a StoryActivityWatcher. Whenever a module calls
ModuleContext.Active(), StoryActivityWatcher.OnStoryActivity() is
called with the story ID of the module's story.

This mechanism gives the user shell an indication of what stories
have ongoing activity. This can be used to avoid calling the
screen saver over a playing youtube video. In the future, ongoing
activity may be established from other sources than
ModuleContext.Active() calls, for example link updates or display
updates, it may involve more information (such as an expected
duration of an activity), and it may establish policies and
decisions made in the framework. See the discussion on
StoryActivityWatcher for some arguments why we don't needs this
information right now, and why we don't unduly constrain ourselves
by not processing it yet.

Test this end to end between a module and a user shell in the
user_shell integration test. Add a new common_active_module that
keeps calling Active() every second.

TESTED=/system/test/run_modular_tests.sh user_shell

MI4-1011 #done

Change-Id: I93ac68a7f79517681a183427cd063cb6b1b62dda