[rust] Add option to enable lto for a rust binary
A rustc_binary can set with_lto = "none|thin|fat" to force LTO to be
applied (or not) to the target. A GN arg rust_lto takes the same values
to set LTO for all rustc_binaries that don't set with_lto. If neither
with_lto nor rust_lto are set, it is disabled for debug builds and set
to "fat" for release builds.

TC-148

Change-Id: Ia735e6841232b304e6a380e15230076e1420bed6