[build] error out duplicate go dependencies
build.py can produce confusing errors if a user accidentally duplicates
a dependency. Before this patch, build.py would first create a proper
symlink for the first dependency, but the second dependency will try to
make a symlink inside the second dependency, which will cause a failure
since the real path then won't be inside the build directory. This patch
fixes this by raising an exception if any dependencies target the same
destination.

Change-Id: If49fc301c3a916b16d8beebdb675a3893e2fd624