[sysmgr] Add integration test for system service startup.
This integration test consists of a (1) a test binary, (2) a dummy
interface and (3) a dummy service that implements the dummy interface.

The test binary includes appmgr as a library to create a hermetic
environment for testing. It then passes a test config to sysmgr via
command line flags. It then connects to the service at
svc/sysmgr.test.Interface and waits for a pre-defined response
The test binary links against gtest for ease of reporting.

CP-38 #done
CP-55 #progress

Change-Id: Ida825cfedce2cca48a51e4dd2b0321a51822759c