[drivers][optee] Add barebones OP-TEE driver
OP-TEE is a Trusted OS designed for executing in the secure world of
Cortex-A cores using TrustZone. Interaction with OP-TEE is an RPC like
interface using the ARM Secure Monitor Call. The OP-TEE OS (BL32) is
initialized by ARM Trusted Firmware (BL31) prior to U-Boot (BL33).

This commit adds a not yet useful barebones driver that does nothing
more than check to see whether OP-TEE is available in the secure world.
For now, the platform driver will only attempt to initialize OP-TEE on
VIM2.

Test: Manually tested on VIM2 that the driver starts and binds when
OP-TEE is installed and that it doesn't when it is not installed.

Change-Id: I15315191d1a66f6841ec66e9bd7e7e8739aaa29e