[x86] Fix AVX register decode.
Fixes a kernel panic if the x86 vector registers were set or retrieved
and the CPU supports the compressed xsave format.

The saved state_component array was previously indexed by
  component - 2
which confused me in a previous patch. This change converts the
array to be indexed by component and just leaves the 0- and 1-index
entries empty.

Adds documentation to this effect and re-enables the tests.

ZX-2203

Change-Id: I7350d8230de22d947bbf355efc35c6f6bdca4fbb