[Codec] codec_android_pal to build AAC decoder
This target is to allow the AAC decoder (whose BUILD.gn
is in a later CL) to build for and run on Fuchsia.

The omx_android_pal is larger and will be in a later CL.
They're not the same target since this target can be
much smaller and this target can be focused on just
building the core AAC lib, not the OMX stuff.

Change-Id: Iaf36459c3ad689e93c8fe608e020861f9d709da2