[ledger] Unify some default benchmark parameters.
Make most benchmarks operate on the same default parameters (size of the
key and value, number of entries), and give descriptive names to the
ones that are different (such as *_big_entry).
Update entry benchmark now have a "--value-size" parameter" to match the
rest of the local modifications benchmarks.
Also get rid of unnecessary splits (to make output more readable on
dashboards).

LE-486 #done

Change-Id: I886dcfaf7e0fa68fb0fc2f51dfcfb5ec93706a2e