[disk-pave] Add support for ARM platforms to paver
A new device-partitioner implementation was added that doesn't expect a
GPT was added. It does not have the ability to remove partitions or add
them, but rather only find them.

In addition, the new paver implementation works with the ZIRCON-A/B/R
partition layout instead of EFI/KERN-C partitions that ChromeOS and EFI
based systems expect.

Tested:
Manually tested paving on pixelbook and vim2.

Change-Id: I49005a8376e00fa034bfc77b7c0ffe752712207b