[disk-pave] Validate ZBI payloads.
Support was added to validate ZBI payloads to ensure ARM payloads do not
end up on x86 devices and vice-versa. This only works on ARM platforms
today as other platforms have not switched over ZBI based partiitons.

This requires streaming the entire ZBI to the VMO instead of in chunks
as well as writing it to disk in one shot. This is not much of a
performance loss since Zircon is fairly small compared to the FVM which
takes the majority of the paving time. This could have been skipped
since the first chunk in the VMO contains everything we need to validate
(the first two headers) but it is error prone to future changes
potentially.

Tested:
Manually tested on VIM2 and pixelbook.

Change-Id: Ic874eb6c9c8a095a7763d82da4840895be1ce215