[debugger] Index source files.
Creates an index of source file names containing code. This allows
efficient conversion from file name to DWARF compile unit. It is not
currently hooked up but in the future will support file:line
breakpoints.

Clear compile units as we process them. For a release build of Chrome
this changes it from 11.2GB to 5.7GB. This requires storing the offset to the
DIE rather than the actual DIE object in the index.

Add an #ifdefed-out benchmark for symbol load performance. I run this on
Chrome to manually check my changes, although this isn't something that
can be enabled by default.

Add some more _util files for random helper functions. Rename
host_utils to host_util for consistency.

Change-Id: I50eaf006c90221d148f03dd85c5921b1bb3369b3