[images] Ensure that pkgsvr is always included
It is all too often that users produce incomplete package manifests,
resulting in a boot failure at blob-init, because the build system includes
the command line argument to pkgsvr, but not pkgsvr itself. This adds pkgsvr,
which should result in a failure later in the boot process, for example
explicitly stating that appmgr is missing.

Change-Id: Iad4b4ee25c99ba69d12102e4cac95ead615063f5