[dev][intel-hda][dsp] poll the CPA bit for power up/down
The driver sets the SPA bit to power up/down the DSP cores
and polls the CPA bit to match the SPA value. When the
values match the power transition has been achieved.

Previously it was polling the wrong bit so the DSP may
be in the wrong state for subsequent operations.

Change-Id: I6ce9a3fa7bfcf9978b443da732e03da33e0201f2