[Codec] Build AOSP OMX SW AAC decoder for Fuchsia
This uses omx_android_pal to build the AOSP OMX SW AAC
decoder for Fuchsia.  The generated .so loadable_module
has a C-only ABI (part of omx_android_pal adapts from
the C-only ABI to the C++ ABI that AOSP uses while keeping
the C++ ABI purely within the .so).

Change-Id: Ifb9289f41af9deecc93d11a980e38436f5657200