[Codec] use_aac_decoder
The use_aac_decoder() code uses CodecFactory to get a Codec and decode
AAC ADTS data from a file.

The use_aac_decoder example can be used to decode to an output WAV file,
or to just calculate the sha256.

The use_aac_decoder_test test decodes the very short test_audio.adts
recording (which has matching LICENSE) faster than real-time and checks
the SHA256 against a golden SHA256 from a manual decode that sounded
correct (unfortunately one golden sha256 for x64 and one for amd64 for
now - both sound correct but the low-order bit is different by 1 in a
relatively low number of bytes - TODO on that for now).

In use_aac_decoder_test, the CodecFactory instance is created using a
local appmgr which creates its own sysmgr which creates the CodecFactory
process, but asking the local sysmgr for CodecFactory isn't using the
normal mechanism yet.  The current way is mostly from this commit:
https://fuchsia.googlesource.com/garnet/+/dd4cff2088c60d1c74a2f76ab9f4c.

The example and test are two separate binaries, and are in separate
packages; they both call use_aac_decoder().

Change-Id: Ie9106a6d35a199b9966c3320b7cffc535557fa93