[machina][vsock] Remove writable_ set.
Since we can't control what commands are sent from the guest to the
device we can elide this tracking on the host side. Instead we should
just poll the TX queue and handle the messages based on the current
connection state.

This fixes some issues when guests connect to a host port.

Change-Id: I049fbfcbee546aad894188cbff3dbfde87ad4f19