[ledger] Fix flake in PageUpload
It is possible for two OnNewCommit to arrive close to one another, one
for a diverging commit (multiple heads), and one for the merge. In that
case, it is possible to end up in a state where we think we need to wait
for the merge commit while it already has happened, and we crash.

We introduce even more state in PageUpload so that at most one
processing stream is executed at any time.

LE-458 #done

Change-Id: Ied8c405f81fad239932365f38e2d6bff754acd5b