[pkg] Template rule for packages with test binaries.
We do encourage multiple separate test *binaries* even
for tests from the same directory, because each binary
gets its own entry on CI status page, and has a separate
log file, and thus improves visibility of test results.

At the same time, we want as few as possible *packages*
per directory, because every package we define in a
BUILD file must be referenced in some file over in the
.../packages hierarchy in order to get included in the
build.

Therefore, we add a rule template tests_package() that
abbreviates the declaration of the tests = [] from the
deps = [], but we don't enforce a 1:1 correspondence
of executable() and package() as we do with
executable_package() for prod binaries.

MI4-1018 #comment  Template rule for packages with test binaries.

TESTED=runtests /pkgfs/packages/*/0/test

Change-Id: I95ac976310ce72dc10ab12c53ebfcc060bab92d1