[amber] Fix fetching blobs if there is a network error
We had a bug where if the server was offline, amber would
dereference a nil pointer. This happened because we were
capturing the error in an `if` scope, but returning an old
`err` from the outer scope. This fixes the bug by lifting
the `client.Get` call into the outer scope so we return
the proper error.

Change-Id: I77f0c4ddf0970e04af897f9253076b2381835f9f