[cleanup][future] Add another Operation wrapper: NewCallbackOperation.
The existing WrapFutureAsOperation is hard to grok. This simplifies it a
little bit by passing a callback that runs the operation and returns a
FuturePtr that can be consumed to be notified of the completion of the
operation.

An OperationBase* is given to the callback such that it can guard any
portion of the async operation using a WeakPtr that is tied to the
operation instance, and thus indirectly to the OperationContainer in
which it lives.

Both are marked experimental while we explore how best to inline
operation bodies.

Change-Id: I980115c065cbfb3d6285baa38bb0cd1a397800f2