[fit] Port most services to fit::function callbacks.
Note, I had to disable the VsockServerTest::HostConnectFreeEphemeralPort
unit test.  It seems that it was only passing before due to a dangling
pointer and now it fails after switching to fit::function.  I was
able to reproduce the problem on the prior version after fixing the
memory error upon which the test relied.  Filed PD-155.

Change-Id: Ibea36ff723a1a8fe5595c72ef3edf5b5c6e61056