[debugger] Allow breakpoitns on file:line.
Adds the ability to look up the addresses for file:line combinations in
the symbols system. There is some testing added for this.

Hooked up breakpoints to this code to support breakpoints on file:line
locations.

Exposed a breakpoint location object through the client API. This will
need to be surfaced in the UI like in other debuggers. The UI is not
currently hooked up.

Disables the unit memory optimization. A TODO is added to fix this as
it turns out this needs more work.

Change-Id: I40d6098deee19bb020b289953674f626ab621cae