[fidl] Use loop.Run() rather than WaitForResponse() in compat test
InterfacePtr is not safe to use from multiple threads simultaneously.
Having more than one thread on the event loop was leading to races: one
thread was blocked on WaitForResponse() but the other thread actually
received the response.

Also re-enable the test since it's no longer flaky.

Tested: by running it 100 times in QEmu with KVM disabled.
Bug: FIDL-214 #done

Change-Id: Ie3f5b14d2d5e9262931d048f14e03467b0e44c09