[userboot] Stop secondary CPUs before shutting down
ZX_SYSTEM_POWERCTL_SHUTDOWN doesn't cleanly shutdown the system. It
appears that it can halt secondary CPUs while they are still holding
locks that the primary may later attempt to acquire. This can lean to
the system hanging on shutdown.

This change is a temporary mitigation.

I repro'd the hang behavior on x64 with qemu+kvm.

I tested this change on x64 with qemu+kvm and on arm64 with qemu by
booting the system in a loop repeatedly with a hacked up userboot
program that called do_shutdown.  I observed that without this change
it would hang about 1 out of 10 times. With the change, I observed no
hangs in several hundred cycles.

INTK-390

Change-Id: I2631908a2f462372a6ec726b3b6b3437fcac8de9