[fidl] Prepare for WaitForResponse deprecation
Now that FIDL supports events, WaitForResponse and WaitForResponseUntil
are footguns because the next message that comes out of the channel
might be an event rather than a response to the message you just sent.

We're going to remove these methods from InterfacePtr. Clients should
migrate to InterfaceSyncPtr or, in the case of tests, to spinning their
async::Loop until the proper message arrives (or UntilIdle).

This CL introduces an #ifdef that whitelists the existing clients. Once
we work through migrating the clients to more robust APIs, we'll remove
the methods.

Change-Id: I7e9313d0bf66ef480ebc1ecfa366be8ea23be78e