[mediacodec] rename media_codec namespace to fuchsia.mediacodec
This CL changes the media codec FIDL files and C++ include references
so the namespace is now fuchsia.mediacodec. This will get us past the
new gate we intend to apply that breaks the build if our FIDL
namespaces don't start with "fuchsia.".

Currently, the implementation of mediacodec isn't part of the build.
This CL doesn't modify the implementation (other than to change the
includes), so the implementation (but not the build) is certainly
broken.

Change-Id: I1d62ff7c251a434e994a4ea4c21c22e33dc00903